Add changes to project start
Rename package Implementations to implementations;
Change type User to Optional<User> findByEmail(String email);
