+++ src/main/java/readerblog/mates/readerblog/controllers/AuthorAdminController.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import readerblog.mates.readerblog.entities.Author;
-import readerblog.mates.readerblog.services.Implementations.AuthorAdminService;
+import readerblog.mates.readerblog.services.implementations.AuthorAdminService;
 
 @Controller
 @RequestMapping("/authorsAdmin")
