+++ src/main/java/readerblog/mates/readerblog/repositories/UserRepository.java
@@ -25,7 +25,7 @@ public interface UserRepository extends JpaRepository<User,Long> {
 
     boolean existsByEmail(String email);
 
-    User findByEmail(String email);
+    Optional<User> findByEmail(String email);
 
     List<User> findByFirstNameAndLastName(String firstName, String lastName);
