+++ src/main/java/readerblog/mates/readerblog/services/implementations/AuthorAdminService.java
@@ -1,4 +1,4 @@
-package readerblog.mates.readerblog.services.Implementations;
+package readerblog.mates.readerblog.services.implementations;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
